+++ test/package.test.js
@@ -11,9 +11,9 @@ describe('@authnomicon/login', function() {
     it('should have assembly metadata', function() {
       expect(json.assembly.namespace).to.equal('org.authnomicon/login');
       expect(json.assembly.components).to.deep.equal([
-        'locals/flowurl',
+        'prompt',
         'service',
-        'login/http/prompt',
+        'locals/flowurl',
         'identifier/service',
         'password/prompt',
         'password/service',