+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.7] - 2024-01-22
 ### Changed
 - `GET /login` handler checks if `module:@authnomicon/credentials.PasswordStore`
 interface is implemented prior to redirecting to `/login/password`.  If not, the
@@ -75,7 +77,8 @@ it will redirect to `/login/password` as before.
 
 - Initial release.
 
-[Unreleased]: https://github.com/authnomicon/login/compare/v0.0.6...HEAD
+[Unreleased]: https://github.com/authnomicon/login/compare/v0.0.7...HEAD
+[0.0.6]: https://github.com/authnomicon/login/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/authnomicon/login/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/authnomicon/login/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/authnomicon/login/compare/v0.0.3...v0.0.4